online_inference/requirements.txt
scikit-learn==0.24.1
pandas==1.1.5
numpy==1.19.5
fastapi==0.63.0
uvicorn==0.13.4
PyYAML==5.3.1
marshmallow-dataclass==8.4.1